chore: extract optimizer slot math into core/slots
site_optimizer.go mixed the stateful optimizer orchestration with the
arithmetic over per-slot timeseries. The latter needs neither the site nor a
loadpoint, so it was only testable through the surrounding request builder.

Move the pure helpers to core/slots and take their tests along. Site methods,
locking and publishing stay in core, so no site interface is needed.
