+++ core/site_optimizer_test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/evcc-io/evcc/api"
 	"github.com/evcc-io/evcc/core/loadpoint"
 	"github.com/evcc-io/evcc/core/types"
-	"github.com/evcc-io/evcc/tariff"
 	"github.com/evcc-io/evcc/util"
 	"github.com/evcc-io/evcc/util/config"
 	optimizer "github.com/evcc-io/optimizer/client"
@@ -28,30 +27,6 @@ func TestLoadpointProfile(t *testing.T) {
 
 	// expected slots: 0.25 kWh...
 	require.Equal(t, []float64{250, 250, 250, 250, 250, 250, 250, 50}, loadpointProfile(lp, 8))
-}
-
-func TestOptimizerHorizon(t *testing.T) {
-	ts := time.Date(2025, 1, 1, 10, 30, 0, 0, time.Local)
-	horizon := optimizerHorizon(ts)
-
-	// 48h plus end of day
-	assert.Equal(t, time.Date(2025, 1, 3, 23, 59, 59, int(time.Second-time.Nanosecond), time.Local), horizon)
-
-	// before 6:00 the day is not extended
-	assert.Equal(t, time.Date(2025, 1, 3, 5, 30, 0, 0, time.Local),
-		optimizerHorizon(time.Date(2025, 1, 1, 5, 30, 0, 0, time.Local)))
-
-	rates := make(api.Rates, 0, 4*96)
-	for slot := ts.Truncate(tariff.SlotDuration); len(rates) < cap(rates); slot = slot.Add(tariff.SlotDuration) {
-		rates = append(rates, api.Rate{Start: slot, End: slot.Add(tariff.SlotDuration)})
-	}
-
-	// 4 days of slots from 10:30, capped at the last slot of Jan 3rd
-	assert.Equal(t, 246, slotsUntil(rates, horizon, len(rates)))
-	assert.Equal(t, time.Date(2025, 1, 3, 23, 45, 0, 0, time.Local), rates[245].Start)
-
-	// shorter forecast is not extended
-	assert.Equal(t, 8, slotsUntil(rates, horizon, 8))
 }
 
 func TestApplyPrecondition(t *testing.T) {
@@ -128,26 +103,6 @@ func TestLoadpointCurrentAction(t *testing.T) {
 			assert.Equal(t, tc.want, loadpointCurrentAction(lp))
 		})
 	}
-}
-
-func TestAsTimestamps(t *testing.T) {
-	// now is 10 minutes into a 15-minute slot
-	now := time.Date(2025, 1, 1, 12, 10, 0, 0, time.UTC)
-
-	// dt[0]=300 means first event is 300s (5min) before end of current slot
-	// dt[1..] just mark subsequent slot boundaries
-	dt := []int{60 * 5, 60 * 15, 60 * 15}
-
-	got := asTimestamps(dt, now)
-
-	// current slot: 12:00–12:15
-	// first timestamp: 12:15 - 5min = 12:10
-	// subsequent: 12:15, 12:30
-	assert.Equal(t, []time.Time{
-		time.Date(2025, 1, 1, 12, 10, 0, 0, time.UTC),
-		time.Date(2025, 1, 1, 12, 15, 0, 0, time.UTC),
-		time.Date(2025, 1, 1, 12, 30, 0, 0, time.UTC),
-	}, got)
 }
 
 func TestUnmodelledPower(t *testing.T) {
@@ -442,28 +397,6 @@ func TestGridExportLimit(t *testing.T) {
 
 	require.NoError(t, site.SetGridExportLimit(7000))
 	assert.Equal(t, 7000.0, site.GetGridExportLimit())
-}
-
-func TestBlendMeasured(t *testing.T) {
-	slots := []float64{100, 100, 100, 100, 100, 100}
-	blendMeasured(slots, 200, 4)
-	assert.Equal(t, []float64{200, 175, 150, 125, 100, 100}, slots)
-
-	// fewer slots than decay length
-	short := []float32{100, 100}
-	blendMeasured(short, 200, 4)
-	assert.Equal(t, []float32{200, 175}, short)
-}
-
-func TestBlendScale(t *testing.T) {
-	slots := []float32{100, 100, 100, 100, 100, 100}
-	blendScale(slots, 2, 4)
-	assert.Equal(t, []float32{200, 175, 150, 125, 100, 100}, slots)
-
-	// fewer slots than decay length
-	short := []float64{100, 100}
-	blendScale(short, 0.5, 4)
-	assert.Equal(t, []float64{50, 62.5}, short)
 }
 
 func TestCurrentSlotSuggestion(t *testing.T) {
