+++ core/slots/slots.go
@@ -1,0 +1,147 @@
+// Package slots provides arithmetic over per-slot timeseries as used by the optimizer.
+package slots
+
+import (
+	"slices"
+	"time"
+
+	"github.com/evcc-io/evcc/api"
+	"github.com/evcc-io/evcc/tariff"
+	"github.com/jinzhu/now"
+	"github.com/samber/lo"
+	"golang.org/x/exp/constraints"
+)
+
+// MatchSoc returns the end time of the first slot whose soc satisfies fun
+func MatchSoc(ts []float32, fun func(float32) bool) time.Time {
+	for i, soc := range ts {
+		if fun(soc) {
+			// TODO first slot
+			return time.Now().Add(time.Duration(i+1) * tariff.SlotDuration).Round(time.Second)
+		}
+	}
+
+	return time.Time{}
+}
+
+// FromNow strips away any slots before "now".
+// The profile contains 48 15min slots (00:00-23:45) that repeat for multiple days.
+func FromNow(profile []float64) []float64 {
+	firstSlot := int(time.Now().Truncate(tariff.SlotDuration).Sub(now.BeginningOfDay()) / tariff.SlotDuration)
+	return profile[firstSlot:]
+}
+
+// BlendMeasured decays the first slots from the measured value into the forecast.
+// Slot 0 uses the measured value, the forecast takes over from slot decaySlots on.
+func BlendMeasured[T constraints.Float](slots []T, measured T, decaySlots int) {
+	for i := range min(decaySlots, len(slots)) {
+		w := T(decaySlots-i) / T(decaySlots)
+		slots[i] = w*measured + (1-w)*slots[i]
+	}
+}
+
+// BlendScale decays a scale factor towards 1 over the first slots.
+// Slot 0 is scaled by the full factor, from slot decaySlots on it is 1.
+func BlendScale[T constraints.Float](slots []T, scale float64, decaySlots int) {
+	for i := range min(decaySlots, len(slots)) {
+		w := float64(decaySlots-i) / float64(decaySlots)
+		slots[i] = T(float64(slots[i]) * (w*scale + (1 - w)))
+	}
+}
+
+// Prorate adjusts the first slot's energy amount according to remaining duration
+func Prorate[T constraints.Float](slots []T, firstSlotDuration time.Duration) []float32 {
+	// return empty slice instead of nil to make api happy
+	if len(slots) == 0 {
+		return []float32{}
+	}
+
+	res := slices.Clone(slots)
+	res[0] = res[0] * T(firstSlotDuration) / T(tariff.SlotDuration)
+	return lo.Map(res, func(f T, _ int) float32 {
+		return float32(f)
+	})
+}
+
+// CurrentRates returns the tariff's rates with past slots removed
+func CurrentRates(t api.Tariff) api.Rates {
+	if t == nil {
+		return nil
+	}
+
+	rates, err := t.Rates()
+	if err != nil {
+		return nil
+	}
+
+	// filter past slots
+	now := time.Now()
+	return lo.Filter(rates, func(slot api.Rate, _ int) bool {
+		return slot.End.After(now)
+	})
+}
+
+// Horizon limits the hosted optimizer to 48 hours, extended to the end of that day.
+// Before 6:00 the extension would add almost a full day, so it only applies past 6:00.
+func Horizon(t time.Time) time.Time {
+	horizon := t.Add(48 * time.Hour)
+	if t.Hour() < 6 {
+		return horizon
+	}
+	return now.With(horizon).EndOfDay()
+}
+
+// Until limits maxLen to the slots starting before the given horizon
+func Until(rates api.Rates, horizon time.Time, maxLen int) int {
+	if i := slices.IndexFunc(rates[:min(maxLen, len(rates))], func(slot api.Rate) bool {
+		return slot.Start.After(horizon)
+	}); i >= 0 {
+		return i
+	}
+	return maxLen
+}
+
+// TimeSteps returns the duration in seconds of each of the next minLen slots,
+// shortening the first one to the remainder of the current slot
+func TimeSteps(minLen int, now time.Time) []int {
+	res := make([]int, 0, minLen)
+
+	eos := now.Truncate(tariff.SlotDuration).Add(tariff.SlotDuration)
+	if d := eos.Sub(now); d > time.Second && d < tariff.SlotDuration {
+		res = append(res, int(d.Seconds()))
+	}
+
+	for i := len(res); i < minLen; i++ {
+		res = append(res, int(tariff.SlotDuration.Seconds())) // 15min slots
+	}
+
+	return res
+}
+
+// AsTimestamps converts slot durations in seconds to slot start times
+func AsTimestamps(dt []int, now time.Time) []time.Time {
+	res := make([]time.Time, 0, len(dt))
+
+	eos := now.Truncate(tariff.SlotDuration).Add(tariff.SlotDuration)
+	res = append(res, eos.Add(-time.Duration(dt[0])*time.Second))
+
+	for i := range len(dt) - 1 {
+		res = append(res, res[i].Add(time.Duration(dt[i])*time.Second))
+	}
+
+	return res
+}
+
+// ScaleAndPrune scales the rates' values and limits them to maxLen slots
+func ScaleAndPrune(rates api.Rates, scale float64, maxLen int) []float32 {
+	res := make([]float32, 0, maxLen)
+
+	for _, slot := range rates {
+		res = append(res, float32(slot.Value*scale))
+		if len(res) >= maxLen {
+			break
+		}
+	}
+
+	return res
+}
